scripts/requirements.txt
MetaTrader5
supabase
python-dotenv
pandas
